Extract FileHandler
