+++ app/controllers/moirai/translation_files_controller.rb
@@ -1,10 +1,10 @@
 module Moirai
   class TranslationFilesController < ApplicationController
-    before_action :load_file_paths_and_hashes, only: [:index, :show, :create_or_update]
+    before_action :load_file_handler, only: [:index, :show, :create_or_update]
     before_action :set_translation_file, only: [:show, :create_or_update]
 
     def index
-      @files = @file_paths.map do |path|
+      @files = @file_handler.file_paths.map do |path|
         {
           id: Digest::SHA256.hexdigest(path),
           name: File.basename(path),
@@ -14,7 +14,7 @@ module Moirai
     end
 
     def show
-      @translation_keys = parse_file(@decoded_path)
+      @translation_keys = @file_handler.parse_file(@decoded_path)
     end
 
     def create_or_update
@@ -28,7 +28,7 @@ module Moirai
     private
 
     def handle_update(translation)
-      translation_from_file = parse_file(@decoded_path)
+      translation_from_file = @file_handler.parse_file(@decoded_path)
       if translation_from_file[translation.key] == translation_params[:value]
         translation.destroy
         flash.notice = "Translation #{translation.key} was successfully deleted."
@@ -46,7 +46,7 @@ module Moirai
     end
 
     def handle_create
-      translation_from_file = parse_file(@decoded_path)
+      translation_from_file = @file_handler.parse_file(@decoded_path)
       if translation_from_file[translation_params[:key]] == translation_params[:value]
         flash.alert = "Translation #{translation_params[:key]} already exists."
         redirect_to_translation_file(translation_params[:file_path])
@@ -68,7 +68,7 @@ module Moirai
     end
 
     def set_translation_file
-      @file_path = @file_hashes[params[:id]]
+      @file_path = @file_handler.file_hashes[params[:id]]
       @decoded_path = CGI.unescape(@file_path)
     end
 
@@ -76,46 +76,8 @@ module Moirai
       params.require(:translation).permit(:key, :locale, :value, :file_path)
     end
 
-    def load_file_paths_and_hashes
-      load_file_paths
-      generate_file_hashes
-    end
-
-    def load_file_paths
-      i18n_file_paths = I18n.load_path
-      @file_paths = i18n_file_paths.select { |path| path.end_with?(".yml", ".yaml") }
-    end
-
-    def generate_file_hashes
-      @file_hashes = @file_paths.map { |path| [Digest::SHA256.hexdigest(path), path] }.to_h
-    end
-
-    def parse_file(path)
-      yaml_content = YAML.load_file(path)
-      root_key = yaml_content.keys.first
-      flatten_hash(yaml_content[root_key])
-    end
-
-    def flatten_hash(hash, parent_key = "", result = {})
-      hash.each do |key, value|
-        new_key = parent_key.empty? ? key.to_s : "#{parent_key}.#{key}"
-        case value
-        when Hash
-          flatten_hash(value, new_key, result)
-        when Array
-          value.each_with_index do |item, index|
-            array_key = "#{new_key}.#{index}"
-            if item.is_a?(Hash)
-              flatten_hash(item, array_key, result)
-            else
-              result[array_key] = item
-            end
-          end
-        else
-          result[new_key] = value
-        end
-      end
-      result
+    def load_file_handler
+      @file_handler = Moirai::TranslationFileHandler.new
     end
   end
 end
